Create Book Endpoint

+++ src/main/java/com/example/restservice/model/dto/BookDto.java
@@ -13,7 +13,7 @@ public class BookDto {
     private Long id;
     private String title;
     private String author;
-    private int stockLevel;
+    private Long stockLevel;
     private String availability;
     private String category;
 }
